upupup

+++ spring5/sts/GuestBookJspTest/src/main/java/com/example/demo/dto/Guestbook.java
@@ -9,17 +9,18 @@ import javax.persistence.Id;
 
 @Entity
 public class Guestbook {
+	
 	@Id
 	@GeneratedValue(strategy=GenerationType.AUTO)
-	private Integer id;
+	private Long id;
 	private String name;
 	private String content;
 	private Date regdate;
-	
-	public Integer getId() {
+
+	public Long getId() {
 		return id;
 	}
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
 	public String getName() {
